refactor(indicators): per-school registry aggregation + extract StanceIndicator
Extract StanceIndicator/_sign_stance to indicators/stances.py (breaks the
library<->lib_* import cycle). Add calc/ package + 8 empty lib_<school>
modules whose CLASSES/SCHEMA merge into REGISTRY/SCHEMA. Built-ins unchanged.
