+++ subprojects/Parametric-Indicators/indicators/lib_bw.py
@@ -1,0 +1,4 @@
+"""bw-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
