+++ subprojects/Parametric-Indicators/indicators/lib_levels.py
@@ -1,0 +1,4 @@
+"""levels-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
