+++ subprojects/Parametric-Indicators/indicators/lib_ma.py
@@ -1,0 +1,4 @@
+"""ma-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
