+++ subprojects/Parametric-Indicators/indicators/lib_osc.py
@@ -1,0 +1,4 @@
+"""osc-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
