+++ subprojects/Parametric-Indicators/indicators/lib_quant.py
@@ -1,0 +1,4 @@
+"""quant-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
