+++ subprojects/Parametric-Indicators/indicators/lib_trend.py
@@ -1,0 +1,4 @@
+"""trend-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
