+++ subprojects/Parametric-Indicators/indicators/lib_vol.py
@@ -1,0 +1,4 @@
+"""vol-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
