+++ subprojects/Parametric-Indicators/indicators/lib_volume.py
@@ -1,0 +1,4 @@
+"""volume-school indicator classes. CLASSES/SCHEMA merged into library.REGISTRY/SCHEMA by library.py."""
+
+CLASSES = ()
+SCHEMA = {}
